python/requirements.txt
pydot==1.4.2
ruamel.base==1.0.0
ruamel.yaml==0.17.21